Task1/requirements.txt
grpcio>=1.32.0
grpcio-tools>=1.32.0
Flask>=1.1.2
requests>=2.20.0
